use correct LONG type for nni_atomic_flag on win32

+++ src/platform/windows/win_impl.h
@@ -52,7 +52,7 @@ struct nni_plat_cv {
 };
 
 struct nni_atomic_flag {
-	unsigned f;
+	LONG f;
 };
 
 struct nni_atomic_bool {
